requirements.txt
# Bot Team - All Dependencies
# Single requirements file for simplicity
# Install with: pip install -r requirements.txt

# ── Core Web Framework ────────────────────────────────
Flask==3.0.0
gunicorn==21.2.0

# ── Configuration & Environment ───────────────────────
python-dotenv==1.0.0
PyYAML==6.0.1

# ── HTTP & API ─────────────────────────────────────────
requests==2.31.0

# ── Authentication ─────────────────────────────────────
authlib==1.3.0
flask-login==0.6.3

# ── SSH (Sally only) ───────────────────────────────────
paramiko==3.3.1

# ── Google Workspace APIs (Fred, Iris) ─────────────────
google-auth==2.25.2
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ── Zendesk Integration (Zac, Sadie) ──────────────────
zenpy==2.0.40

# ── Browser Automation (Banji) ────────────────────────
playwright==1.40.0

# ── Testing ────────────────────────────────────────────
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0
responses==0.24.1
faker==20.1.0
factory-boy==3.3.0
freezegun==1.4.0
pytest-playwright==0.4.3
